refactor: move model_registry from pkg/ to internal/
model_registry has no external consumers — all callers are controllers/,
internal/orchestrator/, internal/routes/, and cmd/neutree-cli — so it
belongs in L2 (internal) rather than L1 (pkg) under the layered
architecture documented in contributing/architecture.md.

Side-effect: file_based.go's import of internal/nfs becomes intra-layer
(L2 → L2), removing the grandfathered R1 exception in
scripts/check-boundaries.sh (cleared in a follow-up commit).

Pure refactor: only import paths and the mockery config change; no
behavior change. Verified by go vet, go test on every touched package
(model_registry, orchestrator, routes/models, controllers,
cmd/neutree-cli/app/cmd/model), and make lint.
