+++ internal/model_registry/.mockery.yaml
@@ -4,6 +4,6 @@ filename: "mock_{{.InterfaceName | snakecase}}.go"
 outpkg: mocks
 with-expecter: true
 packages:
-  github.com/neutree-ai/neutree/pkg/model_registry:
+  github.com/neutree-ai/neutree/internal/model_registry:
     interfaces:
       ModelRegistry:
