+++ internal/model_registry/hugging_face.go
