+++ internal/model_registry/hugging_face_test.go
