+++ internal/model_registry/model_registry.go
