+++ internal/routes/models/models.go
@@ -15,8 +15,8 @@ import (
 
 	v1 "github.com/neutree-ai/neutree/api/v1"
 	"github.com/neutree-ai/neutree/internal/middleware"
-	"github.com/neutree-ai/neutree/pkg/model_registry"
-	"github.com/neutree-ai/neutree/pkg/model_registry/bentoml"
+	"github.com/neutree-ai/neutree/internal/model_registry"
+	"github.com/neutree-ai/neutree/internal/model_registry/bentoml"
 	"github.com/neutree-ai/neutree/pkg/storage"
 )
