+++ internal/routes/models/models_test.go
@@ -12,8 +12,8 @@ import (
 	"github.com/stretchr/testify/mock"
 
 	v1 "github.com/neutree-ai/neutree/api/v1"
-	"github.com/neutree-ai/neutree/pkg/model_registry"
-	model_registry_mocks "github.com/neutree-ai/neutree/pkg/model_registry/mocks"
+	"github.com/neutree-ai/neutree/internal/model_registry"
+	model_registry_mocks "github.com/neutree-ai/neutree/internal/model_registry/mocks"
 	"github.com/neutree-ai/neutree/pkg/storage"
 	"github.com/neutree-ai/neutree/pkg/storage/mocks"
 )
